+++ NEWS.md
@@ -31,6 +31,7 @@
 - `GWAS()`: Return number of records used for each variant and allele
   frequencies in `rayOLS`.
 - Update citation instructions.
+* Use `inherits(., *)` instead of `class(.) == *` (R4 compat).
 
 
 # BGData 2.1.0